Don't prematurely abort if we can't access a reaped dead process
Currently there is a race when a process died and we try and get
the PID info and faultstat terminates because it can't get the
process information.  Don't treat this as a termination error but
instead skip over the now reaped process.

Thanks to Rob Colclaser for reporting this issue.

Signed-off-by: Colin Ian King <colin.king@canonical.com>